+++ test/protractor/spec/addFormToPage.spec.js
@@ -1,5 +1,6 @@
 var po = require('./addFormToPageTest.po');
 var tp = require('./lforms_testpage.po');
+
 describe('addFormToPage test page', function() {
   it('should have two forms on the page', function() {
     po.openPage();